+++ R/reverse_geo_lite.R
@@ -181,7 +181,7 @@ reverse_geo_lite_single <- function(lat_cap,
   api <- prepare_api_url(nominatim_server, "reverse?")
 
   # Compose url
-  url <- paste0(api, "lat=", lat_cap, "&lon=", long_cap, "&format=json")
+  url <- paste0(api, "lat=", lat_cap, "&lon=", long_cap, "&format=jsonv2")
 
   if (isFALSE(full_results)) {
     url <- paste0(url, "&addressdetails=0")